+++ naas_drivers/__init__.py
@@ -37,11 +37,12 @@ from naas_drivers.tools.naas_auth import NaasAuth
 from naas_drivers.tools.naas_credits import NaasCredits
 from naas_drivers.tools.budgetinsight import BudgetInsight
 from naas_drivers.tools.taggun import Taggun
+from naas_drivers.tools.youtube import Youtube
 
 import requests
 import os
 
-__version__ = "0.77.3"
+__version__ = "0.79.0"
 
 __github_repo = "jupyter-naas/drivers"
 
@@ -90,6 +91,7 @@ thinkific = Thinkific()
 naascredits = NaasCredits()
 budgetinsight = BudgetInsight()
 taggun = Taggun()
+youtube = Youtube()
 
 __doc_url = "https://naas.gitbook.io/drivers/"